+++ CHANGELOG.md
@@ -1,9 +1,10 @@
 # Changelog
 
-## NEXT
+## 1.5.0
 
 - Add option to write introspection to local file during code generation
 - Add option to define scalar types without directives in client-side code generation
+- Update deps
 
 ## 1.4.1